CONSTA-60 Перенос части методов в файлы helpers
