+++ src/BarChart/index.tsx
@@ -1,7 +1,6 @@
 import React from 'react'
 
-import { CoreBarChart, Threshold, TypeColumn, UnitPosition } from '@/_private/components/BarChart'
-import { GroupItem } from '@/_private/components/BarChart/components/Group'
+import { CoreBarChart, Threshold, UnitPosition } from '@/_private/components/BarChart'
 import {
   getGroupsDomain,
   getValuesDomain,
@@ -10,7 +9,11 @@ import {
 import { defaultRenderGroup } from '@/_private/components/BarChart/renders'
 import { FormatValue } from '@/_private/types'
 
-import { transformGroupsToCommonGroups } from './helpers'
+import {
+  getColumnsLengthArray,
+  getMaxNumberGroupsArray,
+  transformGroupsToCommonGroups,
+} from './helpers'
 
 export type Column = number | undefined
 
@@ -65,19 +68,6 @@ export const BarChart: React.FC<Props> = props => {
     threshold: props.threshold,
   })
 
-  const getColumnsLengthArray = (groupsItem: GroupItem[], typeColumn: TypeColumn) => {
-    let columnsLengthArray: number[] = []
-
-    groupsItem.map((group: GroupItem) =>
-      group[typeColumn].map(column => {
-        if (column?.sections?.[0]?.value && column?.sections?.[0]?.value !== undefined) {
-          columnsLengthArray = columnsLengthArray.concat(column.sections[0].value)
-        }
-      })
-    )
-    return columnsLengthArray
-  }
-
   const columnsLengthArray = getColumnsLengthArray(commonGroups, 'columns')
   const reversedColumnsLengthArray = getColumnsLengthArray(commonGroups, 'reversedColumns')
   const maxColumnLength =
@@ -85,16 +75,6 @@ export const BarChart: React.FC<Props> = props => {
   const minReversedColumnLength =
     reversedColumnsLengthArray.length > 0 ? Math.min.apply(null, reversedColumnsLengthArray) : 0
 
-  const getMaxNumberGroupsArray = (groupsItem: GroupItem[]) => {
-    let columnsArray: number[] = []
-
-    groupsItem.map((group: GroupItem) => {
-      if (group?.columns) {
-        columnsArray = columnsLengthArray.concat(group.columns.length)
-      }
-    })
-    return columnsArray
-  }
   const maxNumberGroupsArray = getMaxNumberGroupsArray(commonGroups)
   const maxNumberGroups: number =
     maxNumberGroupsArray.length > 0 ? Math.max.apply(null, maxNumberGroupsArray) : 0
